Struct poem_openapi::OpenApiService

pub struct OpenApiService<T, W: ?Sized> { /* private fields */ }

An OpenAPI service for Poem.

Implementations

impl<T> OpenApiService<T, ()>

pub fn new(api: T, title: impl Into<String>, version: impl Into<String>) -> Self

Create an OpenAPI container.

impl<T, W: ?Sized> OpenApiService<T, W>

pub fn webhooks<W2: ?Sized>(self) -> OpenApiService<T, W2>

Sets the webhooks.

pub fn summary(self, summary: impl Into<String>) -> Self

Sets the summary of the API container.

pub fn description(self, description: impl Into<String>) -> Self

Sets the description of the API container.

pub fn terms_of_service(self, url: impl Into<String>) -> Self

Sets a URL to the Terms of Service for the API.

pub fn server(self, server: impl Into<ServerObject>) -> Self

Appends a server to the API container.

Reference: https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.0.md#serverObject

pub fn contact(self, contact: ContactObject) -> Self

Sets the contact information for the exposed API.

pub fn license(self, license: impl Into<LicenseObject>) -> Self

Sets the license information for the exposed API.

Reference: https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.0.md#license-object

pub fn external_document(
self,
external_document: impl Into<ExternalDocumentObject>
) -> Self

Add a external document object.

Reference: https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.0.md#external-documentation-object

pub fn extra_response_header<HT, H>(self, header: H) -> Self where
HT: Type,
H: Into<ExtraHeader>,

Add extra response header

pub fn extra_request_header<HT, H>(self, header: H) -> Self where
HT: Type,
H: Into<ExtraHeader>,

Add extra request header

pub fn cookie_key(self, key: CookieKey) -> Self

Sets the cookie key.

pub fn spec_endpoint(&self) -> impl Endpoint where
T: OpenApi,
W: Webhook,

Create an endpoint to serve the open api specification as JSON.

pub fn spec_endpoint_yaml(&self) -> impl Endpoint where
T: OpenApi,
W: Webhook,

Create an endpoint to serve the open api specification as YAML.

pub fn spec(&self) -> String where
T: OpenApi,
W: Webhook,

Returns the OAS specification file as JSON.

pub fn spec_yaml(&self) -> String where
T: OpenApi,
W: Webhook,

Returns the OAS specification file as YAML.