Enum poem_openapi::types::MaybeUndefined

pub enum MaybeUndefined<T> {
    Undefined,
    Null,
    Value(T),
}

Similar to Option, but it has three states, undefined, null and x.

Example

use poem::{test::TestClient, IntoEndpoint};
use poem_openapi::{payload::Json, types::MaybeUndefined, Object, OpenApi, OpenApiService};
use tokio::sync::Mutex;
use serde_json::json;

#[derive(Object, Clone, Default)]
struct Resource {
    attr1: Option<i32>,
    attr2: Option<String>,
}

#[derive(Object)]
struct UpdateResourceRequest {
    attr1: MaybeUndefined<i32>,
    attr2: MaybeUndefined<String>,
}

struct Api {
    resource: Mutex<Resource>,
}

#[OpenApi]
impl Api {
    #[oai(path = "/get", method = "get")]
    async fn get_resource(&self) -> Json<Resource> {
        Json(self.resource.lock().await.clone())
    }

    #[oai(path = "/put", method = "put")]
    async fn update_resource(&self, req: Json<UpdateResourceRequest>) {
        let mut resource = self.resource.lock().await;

        match req.0.attr1 {
            MaybeUndefined::Null => resource.attr1 = None,
            MaybeUndefined::Value(value) => resource.attr1 = Some(value),
            MaybeUndefined::Undefined => {}
        }

        match req.0.attr2 {
            MaybeUndefined::Null => resource.attr2 = None,
            MaybeUndefined::Value(value) => resource.attr2 = Some(value),
            MaybeUndefined::Undefined => {}
        }
    }
}

let api_service = OpenApiService::new(
    Api {
        resource: Default::default(),
    },
    "Test",
    "1.0",
);

let cli = TestClient::new(api_service.into_endpoint());

cli.get("/get").send().await.assert_json(json!({"attr1": null, "attr2": null}));

cli.put("/put").body_json(&json!({"attr1": 100i32})).send().await.assert_status_is_ok();
cli.get("/get").send().await.assert_json(json!({"attr1": 100i32, "attr2": null}));

cli.put("/put").body_json(&json!({"attr1": null, "attr2": "abc"})).send().await.assert_status_is_ok();
cli.get("/get").send().await.assert_json(json!({"attr1": null, "attr2": "abc"}));

Variants

Undefined

Undefined

Null

Null

Value(T)

Value

Implementations

impl<T> MaybeUndefined<T>

pub fn from_opt_undefined(value: Option<T>) -> Self

Create a MaybeUndefined<T> from Option<T>, returns MaybeUndefined::Undefined if this value is none.

pub fn from_opt_null(value: Option<T>) -> Self

Create a MaybeUndefined<T> from Option<T>, returns MaybeUndefined::Null if this value is none.

pub const fn is_undefined(&self) -> bool

Returns true if the MaybeUndefined<T> is undefined.

pub const fn is_null(&self) -> bool

Returns true if the MaybeUndefined<T> is null.

pub const fn is_value(&self) -> bool

Returns true if the MaybeUndefined<T> contains value.

pub const fn value(&self) -> Option<&T>

Returns None if the the MaybeUndefined<T> is undefined or null, otherwise returns Some(&T).

pub fn value_mut(&mut self) -> Option<&mut T>

Returns None if the the MaybeUndefined<T> is undefined or null, otherwise returns Some(&mut T).

pub fn take(self) -> Option<T>

Converts the MaybeUndefined<T> to Option<T>.

pub const fn as_ref(&self) -> MaybeUndefined<&T>

Converts from &MaybeUndefined<T> to MaybeUndefined<&T>.

pub const fn as_opt_ref(&self) -> Option<Option<&T>>

Converts the MaybeUndefined<T> to Option<Option<T>>.

pub fn as_opt_deref<U>(&self) -> Option<Option<&U>> where
U: ?Sized,
T: Deref<Target = U>,

Converts the MaybeUndefined<T> to Option<Option<&U>>.

pub fn contains_value<U>(&self, x: &U) -> bool where
U: PartialEq<T>,

Returns true if the MaybeUndefined<T> contains the given value.

pub fn contains<U>(&self, x: &Option<U>) -> bool where
U: PartialEq<T>,

Returns true if the MaybeUndefined<T> contains the given nullable value.

pub fn map<U, F: FnOnce(Option<T>) -> Option<U>>(
self,
f: F
) -> MaybeUndefined<U>

Maps a MaybeUndefined<T> to MaybeUndefined<U> by applying a function to the contained nullable value

pub fn map_value<U, F: FnOnce(T) -> U>(self, f: F) -> MaybeUndefined<U>

Maps a MaybeUndefined<T> to MaybeUndefined<U> by applying a function to the contained value

pub fn update_to(self, value: &mut Option<T>)

Update value if the MaybeUndefined<T> is not undefined.

Example

use poem_openapi::types::MaybeUndefined;

let mut value = None;

MaybeUndefined::Value(10i32).update_to(&mut value);
assert_eq!(value, Some(10));

MaybeUndefined::Undefined.update_to(&mut value);
assert_eq!(value, Some(10));

MaybeUndefined::Null.update_to(&mut value);
assert_eq!(value, None);

impl<T: Deref> MaybeUndefined<T>

pub fn as_deref(&self) -> MaybeUndefined<&T::Target>

Converts from MaybeUndefined<T> (or &MaybeUndefined<T>) to MaybeUndefined<&T::Target>.

impl<T, E> MaybeUndefined<Result<T, E>>

pub fn transpose(self) -> Result<MaybeUndefined<T>, E>

Transposes a MaybeUndefined of a Result into a Result of a MaybeUndefined.

MaybeUndefined::Undefined will be mapped to Ok(MaybeUndefined::Undefined). MaybeUndefined::Null will be mapped to Ok(MaybeUndefined::Null). MaybeUndefined::Value(Ok(_)) and MaybeUndefined::Value(Err(_)) will be mapped to Ok(MaybeUndefined::Value(_)) and Err(_).