Struct poem_openapi::types::Password

pub struct Password(pub String);

A password type.

NOTE: Its type is string and the format is password, and it does not protect the data in the memory.

Tuple Fields
0: String